Introduced CompositeBuildParticipantBuildState interface
This is a follow up, and partial revert, of 7a9aa2a.

The BuildState interface is currently used for too many different
concerns. The interface here is for all usages of a "BuildSate" in the
context of configuring a composite build tree and the task graph(s) of
it. This should be detangled more int the future.